+++ src/js/const/buy-request-statuses.const.js
@@ -1,0 +1,18 @@
+export const BUY_REQUEST_STATUSES = Object.freeze({
+  paid: {
+    labelTranslationId: 'buy-request-statuses.paid',
+    value: 'paid',
+  },
+  rejected: {
+    labelTranslationId: 'buy-request-statuses.rejected',
+    value: 'rejected',
+  },
+  timeout: {
+    labelTranslationId: 'buy-request-statuses.timeout',
+    value: 'timeout',
+  },
+  pending: {
+    labelTranslationId: 'buy-request-statuses.pending',
+    value: 'pending',
+  },
+})