+++ src/vue-router/routes.js
@@ -36,4 +36,7 @@ export const vueRoutes = Object.freeze({
   sponsorshipOutgoingRequests: { name: 'app.sponsorship-outgoing-requests' },
   promoCodes: { name: 'app.promo-codes' },
   promoCodesList: { name: 'app.promo-codes.list' },
+  statistics: { name: 'app.statistics' },
+  statisticsSalesHistory: { name: 'app.statistics-sales-history' },
+  statisticsGeneral: { name: 'app.statistics-general' },
 })