+++ src/vue/fields/MultiSelectField.vue
@@ -1,10 +1,23 @@
 <template>
   <div class="multiselect-field">
-    <div class="multiselect-field__select" @click="toggleDropdown">
-      <span class="multiselect-field__select-label">{{ getLable }}</span>
+    <div
+      class="multiselect-field__select"
+      :class="{ 'multiselect-field__select--active': isDropdownOpen }"
+      @click="toggleDropdown"
+    >
+      <template v-if="label">
+        <label class="multiselect-field__select-label">
+          {{ label }}
+        </label>
+      </template>
+
+      <span class="multiselect-field__select-value-label">
+        {{ valueLabel }}
+      </span>
+
       <i
-        class="multiselect-field__selected-icon mdi mdi-chevron-down"
-        :class="{ 'multiselect-field__selected-icon--active': isDropdownOpen }"
+        class="multiselect-field__select-icon mdi mdi-chevron-down"
+        :class="{ 'multiselect-field__select-icon--active': isDropdownOpen }"
       />
     </div>
 
@@ -61,6 +74,7 @@ export default {
     options: { type: Array, required: true },
     enabledOptionAll: { type: Boolean, default: true },
     selectAll: { type: Boolean, default: false },
+    label: { type: String, default: '' },
   },
   data: _ => ({
     isDropdownOpen: false,
@@ -69,19 +83,19 @@ export default {
   }),
 
   computed: {
-    getLable () {
-      let lable = ''
+    valueLabel () {
+      let value = ''
 
       if (this.enabledOptionAll && this.isSelectedAll) {
-        lable = globalize('multiselect-field.all-selected-lable')
+        value = globalize('multiselect-field.all-selected-value-label')
       } else if (this.selectedOptions.length) {
-        lable = this.selectedOptions.length > 1
+        value = this.selectedOptions.length > 1
           ? `${this.selectedOptions[0].name}, ...`
           : this.selectedOptions[0].name
       } else {
-        lable = globalize('multiselect-field.select-something-lable')
+        value = globalize('multiselect-field.select-something-value-label')
       }
-      return lable
+      return value
     },
 
     isSelectedAll () {
@@ -139,10 +153,24 @@ export default {
   width: 100%;
 }
 
-.multiselect-field__select-label {
+.multiselect-field__select-value-label {
   color: $field-color-text;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 
   @include text-font-sizes;
+}
+
+.multiselect-field__select-label {
+  position: absolute;
+  left: 0;
+  transition: all $field-transition-duration;
+  pointer-events: none;
+  color: $field-color-unfocused;
+  top: 0;
+
+  @include label-font-sizes;
 }
 
 .multiselect-field__select {
@@ -154,10 +182,16 @@ export default {
   width: 100%;
   background: none;
   border: none;
-  padding-right: 2.4rem;
+  padding: 1.5rem 2.4rem 0.6rem 0;
+
+  @include material-border(
+    $field-color-focused,
+    $field-color-unfocused,
+    '&.multiselect-field__select--active'
+  );
 }
 
-.multiselect-field__selected-icon {
+.multiselect-field__select-icon {
   position: absolute;
   right: 0;
   will-change: transform;
@@ -171,7 +205,7 @@ export default {
   }
 }
 
-.multiselect-field__selected-icon--active:before {
+.multiselect-field__select-icon--active:before {
   transform: rotate(-180deg);
 }