+++ CHANGELOG.md
@@ -18,6 +18,7 @@ for further information about branching and tagging conventions.
 - New component 'Card' for displaying all card
 - Skeleton loader for new component 'Card'
 - Global class for markdown `app__markdown`
+- Statistics
 
 #### Removed
 - `card-viewer`