+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
-## [Unreleased]
+## [Unreleased ]
+#### Added
+- `ErrorMessage` component
+
+## [1.3.0-rc.1] - 2019-11-06
 #### Added
 - Trim for input field 
 - Redeem default amount
@@ -14,6 +18,7 @@ for further information about branching and tagging conventions.
 - New component 'Card' for displaying all card
 - Skeleton loader for new component 'Card'
 - Global class for markdown `app__markdown`
+- Statistics
 - Expiration date for assets
 
 #### Removed
@@ -387,7 +392,8 @@ for further information about branching and tagging conventions.
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.3.0-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.3.0-rc.1...HEAD
+[1.3.0-rc.1]: https://github.com/tokend/web-client/compare/1.3.0-rc.0...1.3.0-rc.1
 [1.3.0-rc.0]: https://github.com/tokend/web-client/compare/1.2.0...1.3.0-rc.0
 [1.2.0]: https://github.com/tokend/web-client/compare/1.2.0-rc.0...1.2.0
 [1.2.0-rc.0]: https://github.com/tokend/web-client/compare/1.1.0...1.2.0-rc.0